# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parvath FinServ CRM",
    "version": "1.0.0",
    "description": "All business endpoints require a workspace-scoped session. Mutations require the CSRF token returned by /auth/csrf or /auth/login. Dates use Asia/Kolkata; money is an integer string in paise. See docs/api.md for lifecycle contracts."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "parvath.sid"
      }
    },
...
  },
  "paths": {
...
    "/clients/{id}/documents": {
      "post": {
        "summary": "Upload multipart file into quarantine",
        "tags": [
          "clients"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
    "/documents/{id}/download": {
      "get": {
        "summary": "Authorized controlled download after scan",
        "tags": [
          "documents"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success; JSON {data, meta?}"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Sign-in required"
          },
          "403": {
            "description": "Role or CSRF denied"
          },
          "404": {
            "description": "Record absent or outside workspace"
          },
          "409": {
            "description": "Duplicate, stale version or invalid transition"
          },
          "422": {
            "description": "Zod validation error with fieldErrors"
          },
          "503": {
            "description": "External integration not configured"
          }
        }
      }
    },
...
  }
}